Moved GamedayFrame to separate file
